chore: move `vmm::vstate` code under `vmm::arch`
Move the rest of vstate architecture-specific code under `vmm::arch`.
