--- src/vmm/src/arch/x86_64/vstate/mod.rs
@@ -1,5 +1,0 @@
-// Copyright 2025 Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-/// Architecture specific KVM-related code
-pub mod kvm;
